Move translate editor into Edit/Translate namespace. Also write new test case.
